coredump: Include file mappings
1. Record path for file-backed map
2. Apply the following Fujitsu's fixes
   * POSTK_TEMP_FIX_59
   * POSTK_TEMP_FIX_61
   * POSTK_TEMP_FIX_62
   * POSTK_TEMP_FIX_63

TODO: Dump mappings of ld-*.so and executable as file maps although
they are not mapped in McKernel.

+++ kernel/include/syscall.h
@@ -376,7 +376,7 @@ struct coretable {		/* table entry for a core chunk */
 };
 #else /* POSTK_DEBUG_TEMP_FIX_61 */
 struct coretable {		/* table entry for a core chunk */
-	int len;		/* length of the chunk */
+	off_t len;		/* length of the chunk */
 	unsigned long addr;	/* physical addr of the chunk */
 };
 #endif /* POSTK_DEBUG_TEMP_FIX_61 */
